// models/query.ts
export enum QueryKeys {
  Flight_Search = 'flight-search',
  Flight_Offer_Request = 'flight-offer-request',
  Flight_Offer = 'flight-offer',
  Flight_Places = 'flight-places',
  Flight_Classes = 'flight-classes',
  Flight_Order_Create = 'flight-order-create',
  Flight_Order = 'flight-order',
  Flight_Cancel = 'flight-cancel',
  Flight_Cancellation = 'flight-cancellation',

  // ── eSIM ──────────────────────────────────────────────────
  Esim_Local_Packages = 'esim-local-packages',
  Esim_Global_Packages = 'esim-global-packages',
  Esim_Order_Create = 'esim-order-create',
  Esim_Order = 'esim-order',
  Esim_Sim_Packages = 'esim-sim-packages',
  Esim_TopUp_Packages = 'esim-topup-packages',
}